fix(state): keep snapshot helpers available in sparse checkouts
